+++ tests/api-resources/tenants/templates/versions.test.ts
@@ -1,0 +1,33 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+import Courier from '@trycourier/courier';
+
+const client = new Courier({
+  apiKey: 'My API Key',
+  baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
+});
+
+describe('resource versions', () => {
+  // Prism tests are disabled
+  test.skip('retrieve: only required params', async () => {
+    const responsePromise = client.tenants.templates.versions.retrieve('version', {
+      tenant_id: 'tenant_id',
+      template_id: 'template_id',
+    });
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  // Prism tests are disabled
+  test.skip('retrieve: required and optional params', async () => {
+    const response = await client.tenants.templates.versions.retrieve('version', {
+      tenant_id: 'tenant_id',
+      template_id: 'template_id',
+    });
+  });
+});